# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/market/update":{"post":{"summary":"Update Market","operationId":"update_market_market_update_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MarketUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/signal":{"get":{"summary":"Latest Signal","operationId":"latest_signal_signal_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/market/history/{asset}":{"get":{"summary":"Market History","operationId":"market_history_market_history__asset__get","parameters":[{"name":"asset","in":"path","required":true,"schema":{"type":"string","title":"Asset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/trade/state":{"get":{"summary":"Trade State Status","operationId":"trade_state_status_trade_state_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/candles/{asset}":{"get":{"summary":"Get Candles","operationId":"get_candles_candles__asset__get","parameters":[{"name":"asset","in":"path","required":true,"schema":{"type":"string","title":"Asset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/trade/statistics":{"get":{"summary":"Trade Statistics","operationId":"trade_statistics_trade_statistics_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/market/select/{asset}":{"get":{"summary":"Select Asset","operationId":"select_asset_market_select__asset__get","parameters":[{"name":"asset","in":"path","required":true,"schema":{"type":"string","title":"Asset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/trade/today":{"get":{"summary":"Today Session","operationId":"today_session_trade_today_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/trade/latest":{"get":{"summary":"Latest Trade","operationId":"latest_trade_trade_latest_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/trade/all":{"get":{"summary":"All Trades","operationId":"all_trades_trade_all_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/trade/result":{"post":{"summary":"Trade Result","operationId":"trade_result_trade_result_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TradeResult"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/register":{"post":{"tags":["Authentication"],"summary":"Register","operationId":"register_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/login":{"post":{"tags":["Authentication"],"summary":"Login","operationId":"login_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/me":{"get":{"tags":["Authentication"],"summary":"Get Current User","operationId":"get_current_user_auth_me_get","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/logout":{"post":{"tags":["Authentication"],"summary":"Logout","operationId":"logout_auth_logout_post","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/get-signal":{"get":{"summary":"Get Signal","operationId":"get_signal_get_signal_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Signal"}}}}}}}},"components":{"schemas":{"Candle":{"properties":{"timestamp":{"type":"string","title":"Timestamp"},"open":{"type":"number","title":"Open"},"high":{"type":"number","title":"High"},"low":{"type":"number","title":"Low"},"close":{"type":"number","title":"Close"},"volume":{"type":"number","title":"Volume"}},"type":"object","required":["timestamp","open","high","low","close","volume"],"title":"Candle"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"LoginRequest"},"MarketUpdate":{"properties":{"asset":{"type":"string","title":"Asset"},"timeframe":{"type":"string","title":"Timeframe"},"candles":{"items":{"$ref":"#/components/schemas/Candle"},"type":"array","title":"Candles"}},"type":"object","required":["asset","timeframe","candles"],"title":"MarketUpdate"},"RegisterRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"RegisterRequest"},"Signal":{"properties":{"instruction":{"type":"string","title":"Instruction","default":""},"reason":{"type":"string","title":"Reason","default":""},"asset":{"type":"string","title":"Asset"},"timeframe":{"type":"string","title":"Timeframe"},"session":{"type":"string","title":"Session","default":"UNKNOWN"},"bias":{"type":"string","title":"Bias","default":"NONE"},"action":{"type":"string","title":"Action","default":"WAIT"},"confidence":{"type":"number","title":"Confidence"},"probability":{"type":"number","title":"Probability","default":0.0},"trend":{"type":"string","title":"Trend"},"regime":{"type":"string","title":"Regime","default":"UNKNOWN"},"expiration":{"type":"string","title":"Expiration"},"entry_price":{"type":"number","title":"Entry Price"},"timestamp":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Timestamp"},"risk":{"type":"string","title":"Risk"},"grade":{"type":"string","title":"Grade","default":"N/A"},"reasons":{"items":{"type":"string"},"type":"array","title":"Reasons"},"ema_status":{"type":"string","title":"Ema Status","default":"--"},"ema_strength":{"type":"string","title":"Ema Strength","default":"--"},"rsi_status":{"type":"string","title":"Rsi Status","default":"--"},"rsi_strength":{"type":"string","title":"Rsi Strength","default":"--"},"macd_status":{"type":"string","title":"Macd Status","default":"--"},"macd_strength":{"type":"string","title":"Macd Strength","default":"--"},"volume_status":{"type":"string","title":"Volume Status","default":"--"},"volume_strength":{"type":"string","title":"Volume Strength","default":"--"},"structure_status":{"type":"string","title":"Structure Status","default":"--"},"structure_strength":{"type":"string","title":"Structure Strength","default":"--"},"volatility_status":{"type":"string","title":"Volatility Status","default":"--"},"volatility_strength":{"type":"string","title":"Volatility Strength","default":"--"},"support_status":{"type":"string","title":"Support Status","default":"--"},"support_strength":{"type":"string","title":"Support Strength","default":"--"},"liquidity_status":{"type":"string","title":"Liquidity Status","default":"--"},"liquidity_strength":{"type":"string","title":"Liquidity Strength","default":"--"},"market_state":{"type":"string","title":"Market State","default":"WAITING"},"can_enter":{"type":"boolean","title":"Can Enter","default":false},"entry_window":{"type":"integer","title":"Entry Window","default":0},"countdown":{"type":"integer","title":"Countdown","default":0},"trade_status":{"type":"string","title":"Trade Status","default":"IDLE"},"ema_confirmed":{"type":"boolean","title":"Ema Confirmed","default":false},"macd_confirmed":{"type":"boolean","title":"Macd Confirmed","default":false},"rsi_confirmed":{"type":"boolean","title":"Rsi Confirmed","default":false},"structure_confirmed":{"type":"boolean","title":"Structure Confirmed","default":false},"zone_confirmed":{"type":"boolean","title":"Zone Confirmed","default":false},"adx_confirmed":{"type":"boolean","title":"Adx Confirmed","default":false},"atr_confirmed":{"type":"boolean","title":"Atr Confirmed","default":false},"candle_confirmed":{"type":"boolean","title":"Candle Confirmed","default":false},"pullback_confirmed":{"type":"boolean","title":"Pullback Confirmed","default":false},"agreement_score":{"type":"number","title":"Agreement Score","default":0.0},"confirmation_count":{"type":"integer","title":"Confirmation Count","default":0},"confirmation_total":{"type":"integer","title":"Confirmation Total","default":8},"pattern":{"type":"string","title":"Pattern","default":""},"candle_pattern":{"type":"string","title":"Candle Pattern","default":"NONE"},"candle_strength":{"type":"integer","title":"Candle Strength","default":0}},"type":"object","required":["asset","timeframe","confidence","trend","expiration","entry_price","risk"],"title":"Signal"},"TradeResult":{"properties":{"trade_id":{"type":"string","title":"Trade Id"},"exit_price":{"type":"number","title":"Exit Price"}},"type":"object","required":["trade_id","exit_price"],"title":"TradeResult"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}